[Apollo] View change with f isolated non-primaries
This PR introduces a test case where we isolate a subset of replicas and trigger a view change at the same time.
We first make sure that view change correctly completes in the presence of an adversary isolating a minority (f-1) of non-primary replicas.

Then we deactivate the adversary and make sure the previously isolated replicas catch up with the others.