SimpleThreadPool: fix data race
Since `stopped_` is an atomic but not protected by the mutex,
there was a situation where a job was poped from the job queue but was not executed or released.
In the method:
```
bool SimpleThreadPool::load(Job*& outJob) {
  std::unique_lock<std::mutex> ul(queue_lock_);
  queue_cond_.wait(ul, [this] { return !(job_queue_.empty() && !stopped_); });
  if (!stopped_) {                          << stopped_ == false
    outJob = job_queue_.front();
    job_queue_.pop();                       << job is poped
  }
  return !stopped_;                         << stopped_ is already true
}
```

And later in the thread loop:
```
while (load(j)) {                           << load returns false
  execute(j);
  j->release();
  j = nullptr;
}
```
As a result, the job was not executed and its memory not released.