Add new BFT Client support
This new client is intended to be the canonical BFT client for C++
applications. It's expected that in the future other languages will
have their own clients. A future commit will modify the existing Python
client used in Apollo tests to have the same capabilities as this client.

This client introduces support for variably sized quorums and replica
specific information (RSI). Replica specific information is data
specific to each replica returned with a reply, that is excluded from
quorum matching. The RSI for each replica is returned along with the
response to a `Client::send` call. The client API exists in
`bft_client.h`, with the primary functionality being the blocking `send`
function. send is split into 2 different overloads: one for reading and
one for writing. This enables compiler driven typechecks, since reads
and writes support different quorums. Writes only support `Linearizable`
(2F+c+1) and `ByzantineSafe` (F+1) quorums, while reads support both of
those plus `All` and `MofN` quorums.

The four types of quorums are defined in quorum.h and each can be
parameterized by the destinations to send to. If no destinations are
given, then all replicas in the `ClientConfig` struct are used. If
specific destinations are given, they must contain enough replicas to be
safe. For example, in a `ClientConfig` that contains 4 replicas with an
`f_val` of 1, a `LinearizableQuorum` must contain at least 3
unique destinations, and those destinations must be replicas listed in
the ClientConfig. Additionally, all quorums can be converted to MofN
quorums. This conversion is done internally so that the `Matcher`
component does not need to work with multiple quorums. The reason we
don't just pass an MofN quorum is both for legibility, type safety in
`Client::send` calls and to allow reasonable defaults. For example, all
current code is expected to issue writes with `LinearizableQuorums`, and
does not need to fill in the destinations. This mimics SimpleClient.

There are 2 primary threads that communicate with each other in client
operations. The user of the client has a thread that calls
`Client::send`. This thread blocks, waiting for a quorum of matching
replies from the caller. The other thread is the thread running the
`Communication` instance. It sends and receives raw messages to/from the
network. Received messages are translated into `UnmatchedReply` structs
and put on a queue to be sent to the client thread where they can be
matched and the quorum can be checked. The loop for this matching is in
`bft_client.cpp` and utilizes a `Client::wait` call to get replies.

This code base is split up into a few files containing core classes:

 * `bft_client.h/.cpp` - The main API of the client.
 * `base_types.h` - Types used throughout the rest of the
 client
 * `config.h` - All configuration structs. Each of these should
 be self explanatory. The top level configurations are `WriteConfig` and
 `ReadConfig` that get passed to the `Client::send` function, which is
 the main API function of the library.
 * `exception.h` - Exceptions that can be thrown by the library
 * `quorums.h/.cpp` - All quorum structs, validation, and conversion
 *  `seq_num_generator.h` - The mechanism for generating sequence
 numbers used by SimpleClient. It's a separate class/file to emphasize
 that it is optional for users of the client.
 * `matcher.h/.cpp` - Matches a quorum of replies to ensure it can
 return them to the caller of Clientsend.
 * `msg_receiver.h/.cpp` - The implementation of the `IReceiver`
 interface of the communication module. Raw replies received from the
 network are handled in `MsgReceiver::onNewMessage`, converted to a
 message independent format, `UnmatchedReply`, and put onto a queue so
 they can be matched.

There are also unit tests for the MsgReceiver, Matcher, and quorums in
`bft_client_test.cpp`. There are higher level unit tests for the Client
API in `bft_client_api_tests.cpp`. These tests mock out the
communication layer, and utilize `Behaviors` which pretend to be the
replica network and send replies to client requests. A Behavior is just
a callback function that can send replies. The behavior runs in the
context of the `FakeCommunication` thread and therefore must communicate
with the real client by using the MsgReceiver to return replies.