Introduce a sparse merkle tree DB inspection tool
Introduce a sparse merkle tree DB inspection tool that allows fetching
block data and metadata from a RocksDB database. The tool works on the
command line and includes help output for the supported commands. Data
output is encoded in a simple JSON format. The tool is read-only and
never changes the database.

Rationale for this tool is that the sparse merkle storage is not
convenient for inspection using native RocksDB tools. Reason for that is
the complexity of tree storage on top of RocksDB. For example, the tree
has a number of different key types, constructs blocks on-demand and so
on.

Implement using the v2MerkleTree::DBAdapter class.

Add unit tests for the tool and the needed utility functions.

Example usage:
./sparse_merkle_db_inspector /rocksdb-path getGenesisBlockID
{
  "genesisBlockID": "1"
}