Add server side support for ReplicaSpecificInfo
Execution handlers can now return ReplicaSpecificInfo by specifying its
size and embedding it in the reply buffer. The field is already
supported in ClientReplyMsgHeaders and does not have to be used. Any
execution handler may choose to simply ignore it and let it default to
0.

ReplicaSpecificInfo is not currently supported or used for
pre-execution. It is also not currently used in our test binaries.
Nonetheless, the field is mandatory and was added to all usage of the
`execute` callback. A corresponding change will need to be made in the
product, and will be made before this is merged in.

Additionally, an unused code path related to direct proofs, which we do
not intend to support, was removed.