[Apollo] Checkpoint propagation after restarting all replicas
To test backup and restore, we trigger a checkpoint and then restart
all replicas in a random order with 10s delay in-between. We then
verify the first checkpoint persisted upon restart and then trigger
another checkpoint. We make sure the second checkpoint is propagated
to all the replicas and the initial view is stable.