Add control handlers for upgrade + calling to upgrade on super stable checkpoint (#765)
In this PR we have three parts:

More functionality for checkpointInfo
Callbacks interface to IrequestHandler for control handlers and particular for an upgrade.
Identify super stable checkpoint and invoke the upgrade callback if exist
For part 3 we did the following:
We increased the size of the checkpoint log in one and instead of deleting the new lastStableCheckpoint from the log we delete the previous lastStableCheckpoint (if exist). This way we give time for the previous stableCheckpoint to become super stable (and this time is the time between two consecutive stable sequence numbers).