Add support for state transfer complete callbacks
Add support for a subscription mechanism for State Transfer (ST)
on-complete callbacks. Rationale is that multiple parts of the system
might want to know when ST completes so that they can act accordingly to
the newly-received state. Examples of such cases are pruning and time
management, to name a few.

Provide an implementation for BCStateTran only. Implementations for
NullStateTransfer and SimpleStateTran are not provided at that stage as
the use is unclear.

In order to implement, add a generic CallbackRegistry that supports
multiple registrations, deregistration and arbitrary parameters for the
callbacks.

Add an unit test for CallbackRegistry to demonstrate usage and test
common cases.