Add span to addBlock method
Reason:
The method `addBlock` is overridden in the product and its performance is measured by Opentracing.
The initial implementation did not change the base method but used a tricky workaround based on
saving the parent span between the calls of the `execute` method.

After introducing Pre-execution the method `execute` may be called in the context of different threads.
Unfortunately, spans' behavior is undefined while used together with multithreading.

Solution:
The solution is to avoid saving spans and to create them on demand which involves the change in the base API.

Changes:
* Added `SpanWrapper` to the `addBlock` method.
* Minor changes in the Opentracing utilities.