Add an IStateTransfer getter in kvbc::ReplicaImp
Add the kvbc::ReplicaImp::getStateTransfer() getter. Rationale is that
users of kvbc::ReplicaImp might want to interact with state transfer
(ST). For example, the pruning state machine might want to subscribe
to ST-related events.

The current implementation exposes the whole IStateTransfer interface to
higher layers. That is an overkill. Future commits should split the
IStateTransfer interface so that ST internals are not exposed to users.
That will require substantial refactoring.