Checkpoint propagation after restarting all replicas under load
In this test, under a constant time-service like load, we first
trigger a checkpoint. We then stop all replicas in a random order
with a delay in between of view_change_timeout/n, as we don't want
replicas to start a view change while they stop. We then restart
replicas in a random order with a delay of 10s, we start the primary
at the end, so as to avoid view change being triggered while replicas
restart. We verify checkpoint persisted upon restart and then trigger
another checkpoint. We make sure checkpoint is propagated to all the
replicas.