Support for external create client function in Apollo BftTestNetwork
Currently BftTestNetwork works only with UdpClient. However for more
advanced tests one should be able to different types of clients and the
process of their creation and initialisation should be independent from
BftTestNetwork point of view.

For this reason a new callback is introduced to the constructor, new and
existing class methods of BftTestNetwork class - new_client_cb. It
allows the user of the class to supply custom callback, which creates a
compatible BFT client. Inside the BftTestNetwork class this callback is
used each time when a client should be created.

To keep backward compatability, new_client_cb parameter has got a
default value set to None in new and existing class methods. In this
case BftTestNetwork will use default callback implementation which
creates UdpClient.