Modify SkvbcTracker to work against non-empty blockchain
The current implementation of SkvbcTracker assumes that the blockchain
is empty and all read/write requests are sent via its tracker object.
This works well for apollo tests because they work against tester
replica and each test runs against empty vanilla blockchain.

However Apollo can be used for more advanced tests against predeployed
blockchain which is not empty. This patch introduces the concept of
initial block id.

In nutshell the idea is to set initial block id for the SkvbcTracker and
linearizability checks will be performed only for the blocks following
the initial one.