Fix replica calling tryToEnterView without enough view change messages
Replicas if stopped in a random order with a sufficient delay in between
can initiate view changes, but may end up being stopped in inactive view,
without the view change ever stabilizing.
In such a scenario, a replica stopped in an inactive view will also start
in an inactive view. If another replica initiates view change at the same
time the replica in the inactive view restarts, the restarted replica,
upon receiving the ViewChangeMsg may try to enter the new view without
ever receiving 2f + 2c + 1 ViewChangeMsg. This is because for the replica
restarted from the persistent storage, the curView is set as the latest
active view. If a replica was stopped in an inactive view, the curView
is set to the latest active view and not the last known inactive view it
moved into.
So, if a replica was stopped with curView = 1, latest_active_view = 0,
viewIsActive(curView) = False, it will be recovered with curView = 0,
latest_active_view = 0, viewIsActive(curView) = False, which is
inconsistent, and eventually leads to the replica calling the function
tryToEnterView without ever receiving 2f + 2c + 1 ViewChangeMsgs, and fails
at the assert `v > myLatestActiveView` in tryToEnterView().

This flow can be verified by running the backup and restore test -
test_checkpoint_propagation_after_restarting_all_replicas_under_load, with
in-between delay for stopping replicas set to 10 seconds.
The failure is persistent irrespective of the view number in which the
replica was stopped.

To avoid the failure, when the replica is reloaded from the persistent
storage, we update its curView to lastAgreedView + 1, if it was stopped
in an inactive view.

(cherry picked from commit 637e197a1ccf7de22d23ec6ceb044bace7e296d7)