[Pre-execution] Throw an exception if pre-execution returns error code
Our current implementations of IRequestsHandler::execute() return a non-empty response, even in case of execution failure. Errors are reported by the returned status code (0 meaning success).

Currently, the pre-processor does not interpret this status code and practically considers even failed executions successful, and hands them to post-execution to create blocks.

In this PR, we take a conservative approach by crashing the replica in case of non-zero status code in pre-execution.