Fix uninitialized memory issues in Key Store
Make sure we don't call `ReplicaKeyStore::deserializeReplicaKeyStore()`
with an empty buffer. Doing so silently fails and leaves the
deserialized object in an arbitrary state. A number of failures in the
Key Manager test in CI were observed.

Make sure a `repID` member is initialized properly in the Key Manager
ReplicaKeyStore.ser_der test. Reported by valgrind.