Fix uninitialized memory issues in Key Store (#858)
Make sure we don't call `ReplicaKeyStore::deserializeReplicaKeyStore()`
with an empty buffer. Doing so silently fails and leaves the
deserialized object in an arbitrary state. A number of failures in the
Key Manager test in CI were observed - calling `new` with an arbitrary
size, for example.

Ensure `ClusterKeyStore::buffer_` size is not 0 via an assert in the ctor.
Make sure that `ClusterKeyStore` tests set a valid and non-zero size for
the reserved page buffer.

Make sure a `repID` member is initialized properly in the Key Manager
ReplicaKeyStore.ser_der test. Reported by valgrind.