ci(commitlint): introduce commitlint
The commit introduces commitlint-github-action[1] enforcing enforcing conventional commits[2].

Rationale:
Since Concord-BFT is an open-source project, the commit message has to be well-formatted and self-descriptive.
The commit message should be structured as follows:
```
<type>[optional scope]: <description>

[optional body]

[optional footer(s)]
```

The linter is to be run on pushes and pull requests and will be required.
In case the linter fails, it outputs an explanation and a link to the set of rules.
The set of rules is defined in commitlint.config.js and based on conventional commits[3].
For more information about Conventional Commits please refer to the official web site[4].

[1] - https://github.com/wagoid/commitlint-github-action
[2] - https://conventionalcommits.org/
[3] - https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional#rules
[4] - https://www.conventionalcommits.org/en/v1.0.0/