chore: replace exit(1) with std::terminate
Rationale:
Exit code 1 usually means that the process has failed in a controlled way.
For example: `grep -r "non-existing string" .` returns 1 which
means that it failed to find the string but did not crash.
In Concord-BFT `exit(1)` is mostly used when an unexpected exception is caught and the library has to crash.
`std::terminate` crashes the process with `SIGABRT` and generates a core dump.
However, if a user of Concord-BFT would like to change the behavior they can set a custom terminate handler[1].

[1] - https://en.cppreference.com/w/cpp/error/set_terminate