Add clean metadata functionality
Some reconfiguration actions requires the replica to be reloaded with a fresh new configuration/metadata  and to have only the state (for example, key rotation and adding/removing nodes). For that we introduce a new functionality that lets a replica to decide before it goes down if to comeback with a new metadata storage.