CMFC: Revert static function declarations
Previously, the idea was to generate a header-only source file. In practice,
this will lead to some integration difficulties and hence we decided to use
the generate code for a library instead. The user has the choice to compile
the source files with their own code or separate it out into a library.