Decouple RO replica test from minio code in apollo
In test_skvbc_ro_replica, the test code and the minio management code
(responsible for starting and stopping the service) were too tightly
coupled. This makes the integration of external S3 service (e.g. minio
running in Docker) impossible.

The purpose of this patch is to separate these two logics. The
setUpClass() method of the test accepts optional parameter, which
represents abstraction for the S3 service. If this parameter is not
provided - a default one is used.

The abstraction of the S3 service is nothing more than a class with four
methods:
* optional __init__ method for initialisation
* start() - starts the server
* stop() - stops the server
* cleanup() - performs cleanup in the end of the test

The default implementation of the S3 service is called MinioProcess
which is the logic used up to now - local process which is forked/killed
by the test code.