CMF: Add support for python codegen
Generating python source code from CMF files is now supported. Example code can
be generated with the following command:

```
cd messages/compiler
./cmfc.py --input ../example.cmf --output example --language python
```

Tests can be run like so:

```
cd messages/compiler/python
pytest -q test_pygen.py
```

A future commit will hook the python and cpp tests into our cmake build.
This requires updating install_deps.sh and pushing a new version of the
build container to dockerhub.

The code generation for python is similar to that for C++ in that a
visitor is implemented that constructs code as the AST is walked. The
primary difference is that python is dynamically typed and therefore all
the serialization logic itself is implemented in serialize.py.

CMFSerializer and CMFDeserializer implement recursive serialization
functions that take a value and a list of "serializers". These are
strings or tuples that instruct the serialize and deserialize methods
how to recursively encode/decode the corresponding value. The
serializer lists are constructed by the visitor, thus making the visitor
code very simple.

One other thing to note is that due to the dynamic typing of python,
oneofs and optionals are implemented by simply setting the field to the
corresponding type or None in case of an empty optional.

This code also includes a bugfix for C++ that will ensure that booleans
can only be deserialized if they are a 0 or a 1.