Implement a basic functionality for add/remove node feature + a basic demo in Apollo (#877)
To support adding and removing a node from the cluster we take the following actions:

stop the replica on the very same checkpoint (wedge)
kill all replicas
load the replicas with the new configuration that contains/not-contains the nodes' details
restart the new replicas.
Effectively, this means we start a new blockchain with the state of the former blockchain.
Yet, as concord-bft saves all the configuration data in persistent storage, the first step would be to implement functionality to clear the metadata storage before shutting down a replica (see #876).

In this PR we introduce a working example for removing that metadata storage.

Edit: We also added a n-f wedge support. Notice that currently, this is not safe to use it in other reconfiguration action except for removing nodes.