Add an abstract channel type and bounded mpsc impl
In order to allow separation of functionality into various threads, and
allow message passing across those threads, an abstract channel
interface was created in channel.h. The rationale for the design is
given in comments in channel.h. Please read that. It's important to call
out here that the design is reflective of expectations of communication
patterns among our components, such that we must deal with cyclic,
transitive flows, and cannot allow any sender to block arbitrarily.

A lock based, bounded MPSC channel implementation was created in
mpsc_bounded_channel.h. This is useful for an initial implemnentation,
so that we can get started, but also useful to serve as an example of
how to implement the interface defined in the NullChannel in channel.h.

It's expected that the interface can be implemented using any queue,
including non-blocking ones such as boost::lockfree::spsc_queue.

A test is provided that shows how to use the BoundedMpscSender and
BoundedMpscReceiver types that compose a BoundedMpscChannel.