Path prefix for S3 object store is set as configuration parameter
Path prefix is additional key (consider it something like a directory)
created in the target S3 bucket.
This patch changes its creation from random value to value specified in
the S3 configuration. The prefix can't be a random value because all
fetched state in S3 is lost on RO replica restart.