Diagnostics: Move Status code to new class
Note: There are no functional changes in this code.

In preparation for upcoming changes to add histogram/performance support
to diagnostics, a new StatusHandler class was created. It was split into
a .h and .cpp file such that the diagnostics library is no longer a pure
interface library. This is necessary, because diagnostics will need to
link with hdr_histogram, and we don't want to have to make that a direct
dependency of any binaries if they don't need it directly.