Install a new version of OpenSSL into concord-bft image
The current version 1.1.1 is quite old, from 2018. It has a bug that
prevents running the script create_tls_certs.sh inside a container.
We need to run this script to run Apollo tests with TLS communication.
This version is currently the default when installing using the
official Ubuntu package manager.
Version 1.1.1g fixes that bug and is much recent (and stable).
Here we change install_deps.sh to download, configure, build and install
OpenSSL from https://www.openssl.org.