Add implementation for each subclass of MessageBase to be created from MessageBase. (#916)
* Add implementation for each subclass of MessageBase to be created from MessageBase.
This is needed during deserialization (constructing the messages from the
raw char array which is holding a packed struct). In this case we incorporate a larger
msgBody_ into a MessageBase object, having the header required for MessageBase in the beginning.
We need to later create the concrete object in order to call the accessor methods and
the virtual validate method for the complete msgBody_ (holding the header of the MessageBase
followed by the header of the concrete message followed by the payload).

* Added empty Client request to PrePrepare message used
when we need to fill the current woring window.