Stateless docker build
Rationale:
Avoid restarting the docker container manually on every update of the image.

Solution:
Since there is no intermediate state of the build docker container except the `build` directory, we can
make the docker build stateless. As a result, the update of docker image would be easier: no need to restart the
container after the rebase or to make sure that you’re using the latest image because `docker run` would download the
latest automatically.

Changes:
* Every command launches a docker container: replaced all the `exec` commands with `run`
* `make login` launches a new container only if it does not exist, otherwise it enters the existing
* Forwared the error output of `docker container inspect` to `/dev/null`
* Removed `run-c`, `stop-c`, and `remove-c` targets

Usage:
* Once the commit is merged, need to run `docker stop -t 1 concord-bft`
to allow launching it.
* All the rest is exactly the same