TlsTCPCommunication: Degrade log message
This function is called frequently and the common case is that we already have
a connection with that node. If we don't then there are two other log messages
and one will be printed for sure. Meaning, I don't think we loose any
information at runtime by degrading this message to DEBUG.