Add msg parameter to do_with_retry() in s3 client (#975)
The current implementation of do_with_retry() prints the function name
with typeid(f).name(). However when the function is used with std::bind
the output is unreadable for a human.

This patch adds another msg parameter to do_with_retry() which is a
string used in the log message.