pyclient: Add background re-connect trio tasks
Until now, a connection between a BFT client and replica has been done
in a simplified/naive way, coupled with the Apollo framework. 2 pros:
1) Client would connect/disconnect to a replica using notifications
from the BftTestNetwork.
2) If the replica would disconnect from the client for more than 4
minutes (TCP TIME_WAIT timeout), the client does not know how to
automatically try to re-connect to the replica. New tests might fail.

Here, we remove notifications from the framework and add a background
nursery on the client's creation. The background nursery spawns a task
per destination replica. Each task is responsible to maintain a
connection to the replica. It runs on an infinite loop until
reconnect_flag is False. This happens when the bft_client object
exits (all spawned tasks will end their life).

A task tries to connect to dest_replica on an infinite loop until
informed to quit (on BFT client exit).
There hare 2 states:
1) Connected to dest_replica - in that case, park in the lot.
2) Disconnected from dest_replica - in that case, try to connect to it.
On success, insert the new SSL stream into self.ssl_streams and move to
parking until un-parked. On failure, sleep for 0.1 sec, and retry.

SSL stream might be remove from self.ssl_streams while sending or
receiving data, after finding out that connection is closed or broken.

The background nursery is a general-purpose and may be used for
other background tasks in the future. On production, it is the caller's
responsibility to create and pass it.