take execution in account on vc (#990)
n order to stabilize the viewchange action, we want to take in account the replica progression that is not related to the earliest client request timeout.

For that, we have added another check in the viewchange timer such that if the time from the last execution is less than the viewchange timeout while the time from the earliest pending client request is larger, we double the view change timeout.