Support a client for partitioned DBs
Extend the IDBClient and ITransaction interfaces by providing support
for the notion of `partitions`. A `partition` can be thought of as a key
namespace. Two or more partitions can contain the same key and its value
can be different across them.

In terms of design, the IPartitionedDBClient and the
IPartitionedTransaction interfaces are introduced by extending the
IDBClient and the ITransaction interfaces, respectively. The new
interfaces just add another parameter to the key-value methods, namely
the `partition`. An attempt is made to not change the interfaces
significantly. Moreover, methods for partition management are supported,
i.e. looking up, adding, dropping.

In addition, introduce the PartitionDBClient and PartitionTransaction
classes that bridge partition-based interfaces to the existing IDBClient
one. That way, a partition can be used as drop-in replacement for
existing partition-unaware IDBClient implementations.

Remove the IDBClient::freeIterator() method and make the
IDBClient::getIterator() one return an std::unique_ptr. Remove the
corresponding Guard.

Introduce the IDBClient::startTransaction() method that is an alias to
the IDBClient::beginTransaction() method, but returning an
std::unique_ptr. Support upgrade from the std::unique_ptr to an
ITransaction::Guard in order to automatically commit the transaction.
Applies to IPartitionedTransaction too.

Rename IDBClientITerator::isEnd() to valid() and invert its meaning.
That allows for graceful handling of the case of calling previous() on
an iterator that points to the first element. Fix const-correcness for
the valid() and getStatus() methods.

Move common GTest storage code from `kvbc` to `storage`. Add the
`db_partitions` unit test.

This PR assumes that the underlying store can meaningfully support
partitions. For example, partitions map to column families when RocksDB
is used.

A sample implementation is provided in the existing `memorydb` client.

Finally, this PR is just a proposal and a placeholder for discussions.
Suggestions for changes or totally abandoning the idea are welcome. If
accepted, a RocksDB implementation will be added in a future commit.