Added additional storage to keep information for Sequence Numbers (#996)
* Added additional storage to keep information for Sequence Numbers
for up to N previous working windows. This is an optimisation to
help in the situation where a Replica is catching up and is being
slightly behind the others. If in this case the other Replicas
agree on a stable Checkpoint further than the sequence ID the
behind Replica is requesting, currently we will have to wait for
State Transfer to start to be able to catch up. By keeping the
information for the sequence numbers from the previous Working
Window, the behind Replica will be able to catch up by means of
ReqMissingDataMsg messages.

* Better separation of methods for getting SeqNoInfo from Active Window, History or both.