Support for multiple partitions in multi* methods
Add an ability to have multiple different partitions in multiGet,
multiPut and multiDel methods.

Moreover, since multi* methods in memorydb are implemented trivially by
iterating and calling single-key methods, if an error occurs, then we
assert. Rationale is that the DB may be left in an incosistent state if
we continue.

Make the existing multiGet method const.