Cherry pick v0.8.x into master (#1019)
* Added additional storage to keep information for Sequence Numbers (#996)

* Added additional storage to keep information for Sequence Numbers
for up to N previous working windows. This is an optimisation to
help in the situation where a Replica is catching up and is being
slightly behind the others. If in this case the other Replicas
agree on a stable Checkpoint further than the sequence ID the
behind Replica is requesting, currently we will have to wait for
State Transfer to start to be able to catch up. By keeping the
information for the sequence numbers from the previous Working
Window, the behind Replica will be able to catch up by means of
ReqMissingDataMsg messages.

* Better separation of methods for getting SeqNoInfo from Active Window, History or both.

* Unit tests for SequenceWithActiveWindow and fixes for bugs found by them. (#999)

* Unit tests for SequenceWithActiveWindow and fixes for bugs found by them.

* Third unit test added to move Working Window by its full size
after initial fill of 3 times by half Window.

* Tests require more than 1 Checkpoint.
Otherwise Replicas manage to catch up by means of ReqMissingDataMsg
from the stored history of SequenceNumberInfos in peer Replicas.