Add ExecutionRequest struct
We made this change to introduce the ExecutionRequest struct in the IRequestsHandler interface as part of our work on block accumulation.
This struct includes all the fields that need to be passed to execute() and another field for the execution result.
each request that is represented by this structure will be added to a list when block accumulation is active.