Support Pruning for non-provable keys
This commit adds possibility to mark non-provable keys as
NonProvableStale in case a given key already exists.

Changes:
* The method `addBlock` checks if input non-provable keys already exist
and mark them as stale.
* `DBAdapter::lastReachableBlockKeyDeletes` and `DBAdapter::genesisBlockKeyDeletes` are udpated to support
the cleanup of non-provable keys.
* Added method `DBAdapter::staleIndexKeysForBlock` to look for stale
indices.
* Unit tests.

Additional information:
The non-provable stale keys are stored in the DB in the following way:
```
|EDBKeyType::Key|EKeySubtype::NonProvableStale|Stale since version(block id)|Actual NonProvable key|
```

The non-provable keys are stored in the DB in the following way:
```
|EDBKeyType::Key|EKeySubtype::NonProvable|Actual NonProvable key|Block ID|
```